+++ packages/api/src/brevo-contact/brevo-contact.module.ts
@@ -5,7 +5,7 @@ import { DynamicModule, Module, Type } from "@nestjs/common";
 import { BrevoApiModule } from "../brevo-api/brevo-api.module";
 import { createBrevoContactImportConsole } from "../brevo-contact/brevo-contact-import.console";
 import { BrevoContactImportService } from "../brevo-contact/brevo-contact-import.service";
-import { BrevoEmailImportLogModule } from "../brevo-email-import-log/brevo-email-input-log.module";
+import { BrevoEmailImportLogModule } from "../brevo-email-import-log/brevo-email-import-log.module";
 import { ConfigModule } from "../config/config.module";
 import { TargetGroupInterface } from "../target-group/entity/target-group-entity.factory";
 import { BrevoContactAttributesInterface, EmailCampaignScopeInterface } from "../types";
